Done lab 5

+++ src/Lab5/Domain/Models/AccountData.cs
@@ -4,7 +4,7 @@ namespace Domain.Models;
 
 public class AccountData
 {
-    public AccountData(string? userId, string? userName, string? pincode, decimal balance, Roles? role)
+    public AccountData(string? userId, string? userName, string? pincode, int balance, Roles? role)
     {
         UserId = userId;
         Pincode = pincode;
@@ -15,7 +15,7 @@ public class AccountData
 
     public string? UserId { get; init; }
     public string? Pincode { get; init; }
-    public decimal Balance { get; init; }
+    public int Balance { get; init; }
     public Roles? Role { get; init; }
     public string? UserName { get; init; }
 }
